requirements.txt
# Core FastAPI and Web Framework
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
python-multipart==0.0.6

# Database and Storage
supabase==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Requests and API
requests==2.31.0
httpx==0.25.2

# AI and Machine Learning
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.24.4

# Date and Time Handling
python-dateutil==2.8.2

# JSON and Data Processing
ujson==5.8.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment and Configuration
pydantic-settings==2.1.0

# Async Support
aiofiles==23.2.1

# Type Hints and Validation
typing-extensions==4.8.0

# Additional utilities
email-validator==2.1.0 